requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.2
fake-useragent==1.5.1
frozenlist==1.4.1
greenlet==3.0.3
html2text==2024.2.26
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-core==0.2.11
langchain-text-splitters==0.2.2
langsmith==0.1.84
lxml==5.2.2
multidict==6.0.5
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pdfminer.six==20240706
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-pptx==0.6.23
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.5.0
tiktoken==0.7.0
typing_extensions==4.12.2
urllib3==2.2.2
XlsxWriter==3.2.0
yarl==1.9.4
